requirements.txt
anthropic
python-dotenv
fastapi
uvicorn
yfinance
requests
